requirements.txt
PyYAML==3.11
coverage==3.7.1
nose==1.3.3
wsgiref==0.1.2
